Tests/LibWeb/Text/expected/WebAudio/AudioWorkletNode-parameterData.txt

parameters.size: 2
gain is AudioParam: true
gain.value: 0.80
detune.value: 0
iterated names: gain,detune
PASS
